llm-request-code/pyproject.toml
[tool.poetry]
name = "req-smell-tracing"
version = "0.1.0"
description = ""
authors = ["Redacted"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.33.0"
python-dotenv = "^1.0.1"
pytest = "^8.2.2"
tomlkit = "^0.12.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
